tests: KafkaConfig is used by integration tests only
