+++ tests/integration/utils/config.py
@@ -27,3 +27,20 @@ class KafkaDescription:
     install_dir: Path
     download_url: str
     protocol_version: str
+
+
+@dataclass
+class KafkaConfig:
+    datadir: str
+    kafka_keystore_password: str
+    kafka_port: int
+    zookeeper_port: int
+
+    @staticmethod
+    def from_dict(data: dict) -> "KafkaConfig":
+        return KafkaConfig(
+            data["datadir"],
+            data["kafka_keystore_password"],
+            data["kafka_port"],
+            data["zookeeper_port"],
+        )
