+++ tests/integration/utils/kafka_server.py
@@ -6,9 +6,9 @@ from kafka.errors import LeaderNotAvailableError, NoBrokersAvailable
 from karapace.kafka_rest_apis import KafkaRestAdminClient
 from pathlib import Path
 from subprocess import Popen
-from tests.integration.utils.config import KafkaDescription, ZKConfig
+from tests.integration.utils.config import KafkaConfig, KafkaDescription, ZKConfig
 from tests.integration.utils.process import get_java_process_configuration
-from tests.utils import Expiration, get_random_port, KAFKA_PORT_RANGE, KafkaConfig, KafkaServers, write_ini
+from tests.utils import Expiration, get_random_port, KAFKA_PORT_RANGE, KafkaServers, write_ini
 from typing import Dict, List, Tuple
 
 import logging
